add deployment code to solana directory; add owners
