+++ .github/CODEOWNERS
@@ -147,6 +147,7 @@ core/scripts/gateway @smartcontractkit/dev-services
 /deployment @smartcontractkit/ccip-tooling @smartcontractkit/ccip-offchain @smartcontractkit/keystone @smartcontractkit/core @smartcontractkit/deployment-automation
 /deployment/ccip @smartcontractkit/ccip-tooling @smartcontractkit/ccip-offchain @smartcontractkit/deployment-automation
 /deployment/keystone @smartcontractkit/keystone @smartcontractkit/core @smartcontractkit/deployment-automation
+/deployment/ccip/changeset/solana @smartcontractkit/solana-tooling
 # TODO: As more products add their deployment logic here, add the team as an owner
 
 # CI/CD
